requirements.txt

python-dotenv==1.0.1
supabase==2.13.0
